# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/admin/reports": {
      "get": {
        "operationId": "AdminReportsController_list",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "OPEN",
                "ACTIONED",
                "DISMISSED"
              ],
              "type": "string"
            }
          },
          {
            "name": "targetType",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "STORE",
                "OFFER",
                "RATING"
              ],
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "X-Client-Transport",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "cookie"
              ]
            },
            "description": "Web panels send `cookie` on /auth/otp/verify, /auth/merchant/login, /auth/admin/login, and /auth/refresh to receive the refresh token ONLY as an httpOnly cookie, stripped out of the JSON body. Omit entirely for the mobile app (reads the refresh token from the body)."
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminReportListResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but not permitted (wrong actor type, unapproved merchant, or ownership check failed).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List content reports with takedown-deadline countdown, filterable by status/target type.",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/reports/{id}/action": {
      "post": {
        "operationId": "AdminReportsController_action",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Client-Transport",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "cookie"
              ]
            },
            "description": "Web panels send `cookie` on /auth/otp/verify, /auth/merchant/login, /auth/admin/login, and /auth/refresh to receive the refresh token ONLY as an httpOnly cookie, stripped out of the JSON body. Omit entirely for the mobile app (reads the refresh token from the body)."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminReportActionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentReportDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but not permitted (wrong actor type, unapproved merchant, or ownership check failed).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Act on a report — hides the rating / deactivates the store / cancels the offer, per targetType.",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/reports/{id}/dismiss": {
      "post": {
        "operationId": "AdminReportsController_dismiss",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Client-Transport",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "cookie"
              ]
            },
            "description": "Web panels send `cookie` on /auth/otp/verify, /auth/merchant/login, /auth/admin/login, and /auth/refresh to receive the refresh token ONLY as an httpOnly cookie, stripped out of the JSON body. Omit entirely for the mobile app (reads the refresh token from the body)."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminReportActionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentReportDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated, but not permitted (wrong actor type, unapproved merchant, or ownership check failed).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelopeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Dismiss a report without acting on its target.",
        "tags": [
          "admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "Kurtar API",
    "description": "Kurtar — surplus-food marketplace API. The contract for all four\nfrontends (merchant-web, admin-web, the consumer app, and the\nlanding site).\n\n**Auth**: every route requires a bearer access token\n(`Authorization: Bearer <token>`) unless explicitly marked \"no auth\nrequired\" below — obtained via one of the /auth/* endpoints. A web\npanel additionally sends `X-Client-Transport: cookie` on every\nlogin/verify/refresh call to receive the refresh token ONLY as an\nhttpOnly cookie (never in the JSON body) — the mobile app omits\nthis header and reads the refresh token from the response body\ninstead.\n\n**Errors**: every error response is `{statusCode, errorCode,\nmessage}` (ErrorEnvelope, below) — branch on `errorCode`, never on\n`message` (not stable across releases, and may be Turkish or\nEnglish depending on the request).",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ErrorEnvelopeDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400,
            "description": "HTTP status code, duplicated from the response status for convenience."
          },
          "errorCode": {
            "type": "string",
            "example": "VALIDATION_ERROR",
            "description": "A stable, machine-readable identifier for this specific error condition — the field every client should branch on, never `message`."
          },
          "message": {
            "type": "string",
            "example": "overallStars must be between 1 and 5.",
            "description": "Human-readable detail, Turkish or English depending on the request — not stable across releases, do not pattern-match on it."
          }
        },
        "required": [
          "statusCode",
          "errorCode",
          "message"
        ]
      },
...
      "ContentReportDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "targetType": {
            "type": "string",
            "enum": [
              "STORE",
              "OFFER",
              "RATING"
            ]
          },
          "targetId": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "OPEN",
              "ACTIONED",
              "DISMISSED"
            ]
          },
          "takedownDeadlineAt": {
            "format": "date-time",
            "type": "string"
          },
          "resolvedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "takedownWarningSentAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "takedownBreachedSentAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "targetType",
          "targetId",
          "reason",
          "status",
          "takedownDeadlineAt",
          "createdAt",
          "updatedAt"
        ]
      },
      "AdminReportListItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "targetType": {
            "type": "string",
            "enum": [
              "STORE",
              "OFFER",
              "RATING"
            ]
          },
          "targetId": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "OPEN",
              "ACTIONED",
              "DISMISSED"
            ]
          },
          "takedownDeadlineAt": {
            "format": "date-time",
            "type": "string"
          },
          "resolvedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "takedownWarningSentAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "takedownBreachedSentAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "takedownCountdownMs": {
            "type": "number",
            "description": "Milliseconds until takedownDeadlineAt — negative once breached."
          }
        },
        "required": [
          "id",
          "targetType",
          "targetId",
          "reason",
          "status",
          "takedownDeadlineAt",
          "createdAt",
          "updatedAt",
          "takedownCountdownMs"
        ]
      },
      "AdminReportListResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminReportListItemDto"
            }
          },
          "total": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "pageSize": {
            "type": "number"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "pageSize"
        ]
      },
      "AdminReportActionDto": {
        "type": "object",
        "properties": {
          "note": {
            "type": "string",
            "maxLength": 2000
          }
        }
      },
...
    }
  }
}